area state progress
