+++ src/main/java/game_world/objects/areas/Area.java
@@ -1,10 +1,11 @@
 package game_world.objects.areas;
 
 import game_world.objects.events.Event;
+import interfaces.State;
 
 import java.util.ArrayList;
 
-public abstract class Area {
+public abstract class Area implements State {
 
     /**
      * Main Area class with arbitrary instance attributes
